[CCUBE-1512-1][JY] Add base schema test cases.
