+++ tests/util/translator/generate/elements/email/email-schema-generator.spec.tsx
@@ -20,63 +20,6 @@ describe("EmailSchemaGenerator", () => {
         jest.resetAllMocks();
     });
 
-    it("should generate the base schema WITHOUT validation if required is false", () => {
-        const MOCK_ELEMENT: IEmailField = {
-            label,
-            id: elementId,
-            internalId: "email-field",
-            type: EElementType.EMAIL,
-            required: false,
-            columns: { desktop: 12, tablet: 8, mobile: 4 },
-            validation: [],
-        };
-
-        const generatedSchema =
-            EmailSchemaGenerator.elementToSchema(MOCK_ELEMENT);
-
-        const expectedSchema = generateMockElementSchema({
-            id: elementId,
-            label: {
-                mainLabel: label,
-            },
-            uiType: EElementType.EMAIL,
-        });
-
-        expect(generatedSchema).toStrictEqual(expectedSchema);
-    });
-
-    it("should generate the base schema WITH validation if required is true", () => {
-        const MOCK_ELEMENT: IEmailField = {
-            label,
-            id: elementId,
-            internalId: "email-field",
-            type: EElementType.EMAIL,
-            required: true,
-            requiredErrorMsg,
-            columns: { desktop: 12, tablet: 8, mobile: 4 },
-            validation: [],
-        };
-
-        const generatedSchema =
-            EmailSchemaGenerator.elementToSchema(MOCK_ELEMENT);
-
-        const expectedSchema = generateMockElementSchema({
-            id: elementId,
-            label: {
-                mainLabel: label,
-            },
-            uiType: EElementType.EMAIL,
-            validation: [
-                {
-                    required: true,
-                    errorMessage: requiredErrorMsg,
-                },
-            ],
-        });
-
-        expect(generatedSchema).toStrictEqual(expectedSchema);
-    });
-
     it("should generate the base schema WITH only additional validation if additional validation is added but required is false", () => {
         const validationRule = "@gmail.com";
         const validationErrorMessage =
@@ -118,7 +61,7 @@ describe("EmailSchemaGenerator", () => {
         expect(generatedSchema).toStrictEqual(expectedSchema);
     });
 
-    it("should generate the base schema WITH required & additional validation for if every additional validation is added and required is true", () => {
+    it("should generate the base schema WITH required & additional validation if every additional validation is added and required is true", () => {
         const validationRule = "@gmail.com";
         const validationErrorMessage =
             "This field only accepts gmail addresses";
@@ -157,51 +100,6 @@ describe("EmailSchemaGenerator", () => {
                 {
                     matches: "/^[a-zA-Z0-9._%+-]+@(gmail\\.com)$/",
                     errorMessage: validationErrorMessage,
-                },
-            ],
-        });
-
-        expect(generatedSchema).toStrictEqual(expectedSchema);
-    });
-
-    it("should generate the schema with conditional rendering if conditional rendering is added", () => {
-        const MOCK_ELEMENT: IEmailField = {
-            label,
-            id: elementId,
-            internalId: "email-field",
-            type: EElementType.EMAIL,
-            required: false,
-            columns: { desktop: 12, tablet: 8, mobile: 4 },
-            validation: [],
-            conditionalRendering: [
-                {
-                    fieldKey: "mock456",
-                    comparator: "Equals",
-                    value: "hello",
-                    internalId: "mock456",
-                },
-            ],
-        };
-
-        const generatedSchema =
-            EmailSchemaGenerator.elementToSchema(MOCK_ELEMENT);
-
-        const expectedSchema = generateMockElementSchema({
-            id: elementId,
-            label: {
-                mainLabel: label,
-            },
-            uiType: EElementType.EMAIL,
-            showIf: [
-                {
-                    mock456: [
-                        {
-                            filled: true,
-                        },
-                        {
-                            equals: "hello",
-                        },
-                    ],
                 },
             ],
         });
