+++ tests/util/translator/generate/elements/text/text-schema-generator.spec.tsx
@@ -20,63 +20,6 @@ describe("TextSchemaGenerator", () => {
         jest.resetAllMocks();
     });
 
-    it("should generate the base schema WITHOUT validation if required is false", () => {
-        const MOCK_ELEMENT: ITextField = {
-            label,
-            id: elementId,
-            internalId: "text-field",
-            type: EElementType.TEXT,
-            required: false,
-            columns: { desktop: 12, tablet: 8, mobile: 4 },
-            validation: [],
-        };
-
-        const generatedSchema =
-            TextSchemaGenerator.elementToSchema(MOCK_ELEMENT);
-
-        const expectedSchema = generateMockElementSchema({
-            id: elementId,
-            label: {
-                mainLabel: label,
-            },
-            uiType: EElementType.TEXT,
-        });
-
-        expect(generatedSchema).toStrictEqual(expectedSchema);
-    });
-
-    it("should generate the base schema WITH validation if required is true", () => {
-        const MOCK_ELEMENT: ITextField = {
-            label,
-            id: elementId,
-            internalId: "text-field",
-            type: EElementType.TEXT,
-            required: true,
-            requiredErrorMsg,
-            columns: { desktop: 12, tablet: 8, mobile: 4 },
-            validation: [],
-        };
-
-        const generatedSchema =
-            TextSchemaGenerator.elementToSchema(MOCK_ELEMENT);
-
-        const expectedSchema = generateMockElementSchema({
-            id: elementId,
-            label: {
-                mainLabel: label,
-            },
-            uiType: EElementType.TEXT,
-            validation: [
-                {
-                    required: true,
-                    errorMessage: requiredErrorMsg,
-                },
-            ],
-        });
-
-        expect(generatedSchema).toStrictEqual(expectedSchema);
-    });
-
     it("should generate the base schema WITH only additional validation if additional validation is added but required is false", () => {
         const validationRule = "5";
         const validationErrorMessage =
@@ -183,51 +126,6 @@ describe("TextSchemaGenerator", () => {
                 {
                     matches: validationRegexRule,
                     errorMessage: validationRegexErrorMessage,
-                },
-            ],
-        });
-
-        expect(generatedSchema).toStrictEqual(expectedSchema);
-    });
-
-    it("should generate the schema with conditional rendering if conditional rendering is added", () => {
-        const MOCK_ELEMENT: ITextField = {
-            label,
-            id: elementId,
-            internalId: "text-field",
-            type: EElementType.TEXT,
-            required: false,
-            columns: { desktop: 12, tablet: 8, mobile: 4 },
-            validation: [],
-            conditionalRendering: [
-                {
-                    fieldKey: "mock456",
-                    comparator: "Equals",
-                    value: "hello",
-                    internalId: "mock456",
-                },
-            ],
-        };
-
-        const generatedSchema =
-            TextSchemaGenerator.elementToSchema(MOCK_ELEMENT);
-
-        const expectedSchema = generateMockElementSchema({
-            id: elementId,
-            label: {
-                mainLabel: label,
-            },
-            uiType: EElementType.TEXT,
-            showIf: [
-                {
-                    mock456: [
-                        {
-                            filled: true,
-                        },
-                        {
-                            equals: "hello",
-                        },
-                    ],
                 },
             ],
         });
